requirements.txt
python-telegram-bot
pymongo
dnspython